refactor: jbdcTemlate.update 리턴값 변경

+++ src/main/java/nextstep/session/domain/SessionRepository.java
@@ -3,7 +3,7 @@ package nextstep.session.domain;
 import java.util.Optional;
 
 public interface SessionRepository {
-    int save(Session session);
+    Long save(Session session);
 
     Optional<Session> findById(Long sessionId);
 }
